refactor: implement separation of concerns with Redis service
