+++ backend/user-service/errors/index.js
@@ -1,6 +1,7 @@
 export * from "./auth-errors.js";
 export * from "./user-errors.js";
 export * from "./validation-errors.js";
+export * from "./otp-errors.js";
 
 export {
   AUTH_ERRORS,
@@ -17,3 +18,9 @@ export {
   createValidationErrorResponse,
   sendValidationErrorResponse,
 } from "./validation-errors.js";
+export {
+  OTP_ERROR_MESSAGES,
+  getOTPErrorMessage,
+  createOTPErrorResponse,
+  sendOTPErrorResponse,
+} from "./otp-errors.js";
