+++ backend/user-service/errors/validation-errors.js
@@ -13,10 +13,6 @@ export const VALIDATION_ERRORS = {
 
 /**
  * Helper function to create validation error response
- * @param {Object} error - Error object from VALIDATION_ERRORS
- * @param {Object} details - Validation error details
- * @param {Object} additionalData - Additional data to include in response
- * @returns {Object} Formatted error response
  */
 export function createValidationErrorResponse(
   error,
@@ -38,10 +34,6 @@ export function createValidationErrorResponse(
 
 /**
  * Helper function to send validation error response
- * @param {Object} res - Express response object
- * @param {Object} error - Error object from VALIDATION_ERRORS
- * @param {Object} details - Validation error details
- * @param {Object} additionalData - Additional data to include in response
  */
 export function sendValidationErrorResponse(
   res,
