+++ backend/user-service/server.js
@@ -2,14 +2,14 @@ import http from "http";
 import index from "./index.js";
 import "dotenv/config";
 import { connectToDB } from "./model/user-repository.js";
-import { redisService } from "./services/redis-service.js";
+import { baseRedisService } from "./services/redis/redis-base-service.js";
 
 const port = process.env.PORT || 3001;
 
 const server = http.createServer(index);
 
 // Initialize database and Redis connections
-Promise.all([connectToDB(), redisService.connect()])
+Promise.all([connectToDB(), baseRedisService.connect()])
   .then(() => {
     console.log("MongoDB Connected!");
     console.log("Redis Service initialized!");
